+++ src/main/java/fr/mkadia/mkadiaapi/repositories/CouponRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.Optional;
 
-public interface CouponRepository extends JpaRepository<Coupon, Integer>, JpaSpecificationExecutor<Coupon> {
+public interface CouponRepository extends JpaRepository<Coupon, Long>, JpaSpecificationExecutor<Coupon> {
     Optional<Coupon> findByCode(String code);
     Optional<Coupon> findByCodeAndActiveTrue(String code);
 }
